+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
 import "./App.css";
-import MainPage from "./pages/MainPage/MainPage";
+import MainPage from "./pages/MainPage";
 
 function App() {
   return (
