+++ src/components/RegisterComponent.js
@@ -2,42 +2,7 @@ import { useState } from "react";
 import { Button, Form } from "react-bootstrap";
 import "./Register.css";
 
-const Register = () => {
-  const [id, setId] = useState("");
-  const [username, setUsername] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [samePassword, setSamePassword] = useState(true);
-  const [validated, setValidated] = useState(false);
-
-  const onChangeId = (e) => {
-    setId(e.target.value);
-  };
-  const onChangeUsername = (e) => {
-    setUsername(e.target.value);
-  };
-  const onChangeEmail = (e) => {
-    setEmail(e.target.value);
-  };
-  const onChangePassword = (e) => {
-    setPassword(e.target.value);
-    setSamePassword(e.target.value === confirmPassword);
-  };
-  const onChangeConfirmPassword = (e) => {
-    setConfirmPassword(e.target.value);
-    setSamePassword(e.target.value === password);
-  };
-  const handleSubmit = (e) => {
-    const form = e.currentTarget;
-    if (form.checkValidity() === false) {
-      e.preventDefault();
-      e.stopPropagation();
-      setValidated(true);
-    } else {
-      // TODO: Register API, 중복 체크 API 등등
-    }
-  };
+const RegisterComponent = ({ id, username, email, password, confirmPassword, samePassword, validated, onChangeId, onChangeUsername, onChangeEmail, onChangePassword, onChangeConfirmPassword, handleSubmit }) => {
   return (
     <div className="register">
       <h2 className="register-title">회원가입</h2>
@@ -121,4 +86,4 @@ const Register = () => {
   );
 };
 
-export default Register;
+export default RegisterComponent;
