+++ src/containers/MainPageContainer.js
@@ -1,0 +1,8 @@
+import { React, useState } from 'react';
+import MainPageComponent from '../components/MainPageComponent';
+
+const MainPageContainer = () => {
+    return <MainPageComponent />
+};
+
+export default MainPageContainer;
