+++ src/containers/NonUserMainContainer.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import NonUserMainComponent from '../components/NonUserMainComponent';
+
+const NonUserMainContainer = () => {
+    return <NonUserMainComponent />
+};
+
+export default NonUserMainContainer;
