+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 import "bootstrap/dist/css/bootstrap.min.css";
-import Header from "./components/Header/Header";
+import Header from "./common/Header/Header";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
