+++ src/pages/LoginPage.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import LoginContainer from '../containers/LoginContainer';
+
+const LoginPage = () => {
+    return <LoginContainer />
+};
+
+export default LoginPage;
