+++ src/pages/MainPage.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import MainPageContainer from '../containers/MainPageContainer';
+
+const MainPage = () => {
+    return <MainPageContainer />
+};
+
+export default MainPage;
