+++ src/pages/NonUserMainPage.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import NonUserMainContainer from '../containers/NonUserMainContainer';
+
+const NonUserMainPage = () => {
+    return <NonUserMainContainer />
+};
+
+export default NonUserMainPage;
