+++ src/pages/RegisterPage.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import RegisterContainer from '../containers/RegisterContainer';
+
+const RegisterPage = () => {
+    return <RegisterContainer />
+};
+
+export default RegisterPage;
