--- src/recoil/atom.js
@@ -1,8 +1,0 @@
-// recoil 페이지
-import { atom } from 'recoil';
-
-// for survey
-export const surveyDialogState = atom({
-  key: 'surveyOpen',
-  default: false,
-});
